refactor and tests completed
