+++ client/api/categories.test.js
@@ -21,7 +21,6 @@ describe('API client for categories', () => {
   })
 
   it('fetchCategories returns an error when expected', () => {
-    // expect.assertions(1) // otherwise a valid id would still pass
     return fetchCategories(33).catch(e => {
       expect(e.message).toBe('Error accessing categories api.')
     })
