+++ client/api/fetchFoods.js
@@ -2,7 +2,7 @@ import makeRequest from './requestor'
 
 const foodPath = '/api/v1/foods'
 
-export function fetchFoods () {
+export default function fetchFoods () {
   return makeRequest(foodPath)
     .then(res => res.body)
     .catch(() => { throw new Error('Error accessing categories api.') })
